// src/generated/models/opportunityCreate.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * iPhande API
 * Visibility, opportunity continuity, replay, and grace reflection platform.
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityCreateDescription } from './opportunityCreateDescription';
import type { OpportunityCreateSuburbOrArea } from './opportunityCreateSuburbOrArea';
import type { OpportunityCreatePlaceCode } from './opportunityCreatePlaceCode';
import type { OpportunityCreateLatitude } from './opportunityCreateLatitude';
import type { OpportunityCreateLongitude } from './opportunityCreateLongitude';
import type { OpportunityCreateBudgetAmount } from './opportunityCreateBudgetAmount';
import type { OpportunityCreateImageUrl1 } from './opportunityCreateImageUrl1';
import type { OpportunityCreateImageUrl2 } from './opportunityCreateImageUrl2';
import type { OpportunityCreateExpiryDate } from './opportunityCreateExpiryDate';

export interface OpportunityCreate {
  created_by_profile_id: string;
  title: string;
  description?: OpportunityCreateDescription;
  province: string;
  town_or_city: string;
  suburb_or_area?: OpportunityCreateSuburbOrArea;
  place_code?: OpportunityCreatePlaceCode;
  latitude?: OpportunityCreateLatitude;
  longitude?: OpportunityCreateLongitude;
  category_key: string;
  service_needed: string;
  budget_amount?: OpportunityCreateBudgetAmount;
  contact_name: string;
  contact_phone: string;
  image_url_1?: OpportunityCreateImageUrl1;
  image_url_2?: OpportunityCreateImageUrl2;
  expiry_date?: OpportunityCreateExpiryDate;
}